Memcache(d) server, passive mode

+++ main.go
@@ -3,10 +3,13 @@ package main
 import (
 	"encoding/json"
 	"flag"
+	"net"
 	"net/http"
 	"os"
 	"runtime"
 	"time"
+
+	"github.com/davidmz/ipgeobase/mmc"
 )
 
 const (
@@ -15,12 +18,15 @@ const (
 )
 
 var (
-	dirName        = flag.String("dir", "", "Data directory (required, write access needed)")
+	dirName        = flag.String("dir", "", "Data directory (required, write access needed unless passive mode)")
 	geoBaseURL     = flag.String("url", "http://ipgeobase.ru/files/db/Main/geo_files.zip", "URL of IPGeoBase zip archive")
 	lstAddress     = flag.String("listen", "localhost:7364", "IP address and port to listen")
 	updateInterval = flag.Duration("interval", time.Hour, "Update interval")
+	passiveMode    = flag.Bool("passive", false, "Passive mode: do not write to data directory")
+	debugLevel     = flag.Bool("debug", false, "Debug level log")
 	showHelp       = flag.Bool("help", false, "Show help")
 	showVersion    = flag.Bool("version", false, "Show version number")
+	asMemcache     = flag.Bool("memcache", false, "Serve memcache protocol")
 )
 
 func main() {
@@ -36,6 +42,11 @@ func main() {
 	}
 
 	if !fileExists {
+		if conf.PassiveMode {
+			conf.Log.Error("Database file not exists")
+			os.Exit(1)
+		}
+
 		conf.Log.Info("IPGeoBase file not exists, downloading...")
 		for {
 			os.Remove(conf.Dir.eTagFile())
@@ -54,21 +65,46 @@ func main() {
 		os.Exit(1)
 	}
 
-	go updateRoutine(conf)
+	if conf.PassiveMode {
+		go passiveUpdateRoutine(conf)
+	} else {
+		go updateRoutine(conf)
+	}
 
-	http.HandleFunc("/", func(w http.ResponseWriter, r *http.Request) {
-		ip := r.URL.Query().Get("ip")
-		base := conf.VBase.Load().(*GeoBase)
-		result := base.Find(ip)
-		w.Header().Set("Server", ServerName)
-		w.Header().Set("Access-Control-Allow-Origin", "*")
-		w.Header().Set("Content-Type", "application/json; charset=utf-8")
-		json.NewEncoder(w).Encode(result)
-	})
+	if *asMemcache {
+		conf.Log.Info("Starting memcache server at " + conf.LstAddr.String())
 
-	conf.Log.Info("Starting server at " + conf.LstAddr.String())
-	if err := http.ListenAndServe(conf.LstAddr.String(), nil); err != nil {
-		conf.Log.Errorf("Serve error: %v", err)
-		os.Exit(1)
+		ln, err := net.Listen("tcp", conf.LstAddr.String())
+		if err != nil {
+			conf.Log.Errorf("Serve error: %v", err)
+			os.Exit(1)
+		}
+		h := &MemcacheHandler{conf}
+		for {
+			conn, err := ln.Accept()
+			conf.Log.Debugf("Memcache connect from %q", conn.RemoteAddr())
+			if err != nil {
+				conf.Log.Errorf("Serve error: %v", err)
+				os.Exit(1)
+			}
+			go mmc.NewSession(conn, h)
+		}
+
+	} else {
+		http.HandleFunc("/", func(w http.ResponseWriter, r *http.Request) {
+			ip := r.URL.Query().Get("ip")
+			base := conf.VBase.Load().(*GeoBase)
+			result := base.Find(ip)
+			w.Header().Set("Server", ServerName)
+			w.Header().Set("Access-Control-Allow-Origin", "*")
+			w.Header().Set("Content-Type", "application/json; charset=utf-8")
+			json.NewEncoder(w).Encode(result)
+		})
+
+		conf.Log.Info("Starting HTTP server at " + conf.LstAddr.String())
+		if err := http.ListenAndServe(conf.LstAddr.String(), nil); err != nil {
+			conf.Log.Errorf("Serve error: %v", err)
+			os.Exit(1)
+		}
 	}
 }
